+++ packages/contracts/globals.d.ts
@@ -1,8 +1,5 @@
-declare module 'bn.js';
-declare module 'ethereumjs-abi';
 declare module 'chai-bignumber';
 declare module 'dirty-chai';
-declare module 'yargs';
 
 // HACK: In order to merge the bignumber declaration added by chai-bignumber to the chai Assertion
 // interface we must use `namespace` as the Chai definitelyTyped definition does. Since we otherwise
@@ -29,6 +26,11 @@ declare module 'solc' {
 
 declare module 'web3-eth-abi' {
     export function encodeParameters(typesArray: string[], parameters: any[]): string;
+}
+
+declare module 'ethereumjs-abi' {
+    const soliditySHA3: (argTypes: string[], args: any[]) => Buffer;
+    const methodID: (name: string, types: string[]) => Buffer;
 }
 
 // Truffle injects the following into the global scope