+++ packages/json-schemas/test/schema_test.ts
@@ -23,7 +23,7 @@ const {
     signedOrderSchema,
     signedOrdersSchema,
     blockParamSchema,
-    subscriptionOptsSchema,
+    blockRangeSchema,
     tokenSchema,
     jsNumber,
     txDataSchema,
@@ -170,21 +170,21 @@ describe('Schema', () => {
             validateAgainstSchema(testCases, blockParamSchema, shouldFail);
         });
     });
-    describe('#subscriptionOptsSchema', () => {
+    describe('#blockRangeSchema', () => {
         it('should validate valid subscription opts', () => {
             const testCases = [
                 {fromBlock: 42, toBlock: 'latest'},
                 {fromBlock: 42},
                 {},
             ];
-            validateAgainstSchema(testCases, subscriptionOptsSchema);
+            validateAgainstSchema(testCases, blockRangeSchema);
         });
         it('should fail for invalid subscription opts', () => {
             const testCases = [
                 {fromBlock: '42'},
             ];
             const shouldFail = true;
-            validateAgainstSchema(testCases, subscriptionOptsSchema, shouldFail);
+            validateAgainstSchema(testCases, blockRangeSchema, shouldFail);
         });
     });
     describe('#tokenSchema', () => {