+++ packages/json-schemas/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+v0.x.x - _TBD, 2017_
+------------------------
+    * Rename `subscriptionOptsSchema` to `blockRangeSchema` (#272)
+
 v0.6.7 - _Nov. 14, 2017_
 ------------------------
     * Re-publish JSON-schema previously published under NPM package 0x-json-schemas